requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pillow==10.4.0
pydantic==2.8.2
